+++ models/Member.js
@@ -12,4 +12,20 @@ const memberSchema = new mongoose.Schema(
   { timestamps: true, toJSON: { versionKey: false } }
 );
 
+// Enforce single account holder at the model level.
+memberSchema.pre('save', async function () {
+  if (this.isModified('isBuyer') && this.isBuyer) {
+    await this.constructor.updateMany({ _id: { $ne: this._id } }, { isBuyer: false });
+  }
+});
+
+memberSchema.pre('findOneAndUpdate', async function () {
+  const update = this.getUpdate();
+  const incoming = update?.isBuyer ?? update?.$set?.isBuyer;
+  if (incoming === true) {
+    const doc = await this.model.findOne(this.getQuery(), '_id').lean();
+    await this.model.updateMany({ _id: { $ne: doc?._id } }, { isBuyer: false });
+  }
+});
+
 module.exports = mongoose.model('Member', memberSchema);